refactor before extracting broker-report-parser-api
