+++ src/main/java/ru/investbook/parser/AbstractBrokerReport.java
@@ -19,14 +19,8 @@
 package ru.investbook.parser;
 
 import lombok.*;
-import org.apache.poi.hssf.usermodel.HSSFWorkbook;
-import org.apache.poi.ss.usermodel.Workbook;
-import org.apache.poi.util.CloseIgnoringInputStream;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 import org.spacious_team.table_wrapper.api.ReportPage;
 
-import java.io.IOException;
-import java.io.InputStream;
 import java.nio.file.Path;
 import java.time.Instant;
 import java.time.LocalDate;
@@ -58,19 +52,6 @@ public abstract class AbstractBrokerReport implements BrokerReport {
     private Instant reportEndDateTime;
     @Getter
     private final ZoneId reportZoneId = ZoneId.of("Europe/Moscow");
-
-    protected Workbook getWorkBook(String excelFileName, InputStream is) {
-        try {
-            is = new CloseIgnoringInputStream(is); // HSSFWorkbook() constructor close is
-            if (excelFileName.endsWith(".xls")) {
-                return new HSSFWorkbook(is); // constructor close is
-            } else {
-                return new XSSFWorkbook(is);
-            }
-        } catch (IOException e) {
-            throw new RuntimeException("Не смог открыть excel файл", e);
-        }
-    }
 
     public Instant convertToInstant(String value) {
         value = value.trim();
